split init_rho from main charge.cpp file
