+++ source/module_elecstate/CMakeLists.txt
@@ -16,6 +16,7 @@ list(APPEND objects
     potentials/potential_types.cpp
     potentials/write_pot.cpp
     module_charge/charge.cpp
+    module_charge/charge_init.cpp
     module_charge/charge_broyden.cpp
     module_charge/charge_extra.cpp
     module_charge/charge_mixing.cpp
